+++ src/tablecloth/api.clj
@@ -36,7 +36,7 @@
 
   * `ds` - a dataset
   * `column-name` - if it's existing column name, column will be replaced
-  * `column` - can be column (from other dataset), sequence, single value or function. Too big columns are always trimmed. Too small are cycled or extended with missing values (according to `size-strategy` argument)
+  * `column` - can be column (from other dataset), sequence, single value or function (taking a dataset). Too big columns are always trimmed. Too small are cycled or extended with missing values (according to `size-strategy` argument)
   * `size-strategy` (optional) - when new column is shorter than dataset row count, following strategies are applied:
     - `:cycle` - repeat data
     - `:na` - append missing values
@@ -88,6 +88,8 @@
 
 (defn aggregate-columns
   "Aggregates each column separately"
+  ([ds columns-aggregators]
+  (tablecloth.api.aggregate/aggregate-columns ds columns-aggregators))
   ([ds columns-selector column-aggregators]
   (tablecloth.api.aggregate/aggregate-columns ds columns-selector column-aggregators))
   ([ds columns-selector column-aggregators options]
@@ -102,8 +104,25 @@
 
 
 (defn append
+  "Concats columns of several datasets"
   ([ds & args]
   (apply tablecloth.api.join-concat-ds/append ds args)))
+
+
+(defn array-column->columns
+  "Converts a column of type java array into several columns,
+  one for each element of the array of all rows. The source column is dropped afterwards.
+  The function assumes that arrays in all rows have same type and length and are numeric.
+
+  `ds` Datset to operate on.
+  `src-column` The (array) column to convert
+  `opts` can contain:
+    `prefix` newly created column will get prefix before column number
+  "
+  ([ds src-column opts]
+  (tablecloth.api.join-separate/array-column->columns ds src-column opts))
+  ([ds src-column]
+  (tablecloth.api.join-separate/array-column->columns ds src-column)))
 
 
 (defn as-regular-dataset
@@ -157,6 +176,36 @@
 
 
 (defn column-names
+  "Returns column names, given a selector.
+  Columns-selector can be one of the following:
+
+  * :all keyword - selects all columns
+  * column name - for single column
+  * sequence of column names - for collection of columns
+  * regex - to apply pattern on column names or datatype
+  * filter predicate - to filter column names or datatype
+  * type namespaced keyword for specific datatype or group of datatypes
+
+  Column name can be anything.
+
+column-names function returns names according to columns-selector
+  and optional meta-field. meta-field is one of the following:
+
+  * `:name` (default) - to operate on column names
+  * `:datatype` - to operated on column types
+  * `:all` - if you want to process all metadata
+
+  Datatype groups are:
+
+  * `:type/numerical` - any numerical type
+  * `:type/float` - floating point number (:float32 and :float64)
+  * `:type/integer` - any integer
+  * `:type/datetime` - any datetime type
+
+  If qualified keyword starts with :!type, complement set is used.
+
+
+  "
   ([ds]
   (tablecloth.api.utils/column-names ds))
   ([ds columns-selector]
@@ -177,20 +226,34 @@
   (tablecloth.api.dataset/columns ds result-type)))
 
 
+(defn columns->array-column
+  "Converts several columns to a single column of type array.
+   The src columns are dropped afterwards.
+
+  `ds` Dataset to operate on.
+  `column-selector` anything supported by [[select-columns]]
+  `new-column` new column to create
+  "
+  ([ds column-selector new-column]
+  (tablecloth.api.join-separate/columns->array-column ds column-selector new-column)))
+
+
 (defn complete
   "TidyR complete.
 
-  Fills a dataset with all possible combinations of selected columns. When given combination wasn't existed, missing values are created."
+  Fills a dataset with all possible combinations of selected columns. When a given combination doesn't exist, missing values are created."
   ([ds columns-selector & args]
   (apply tablecloth.api.join-concat-ds/complete ds columns-selector args)))
 
 
 (defn concat
+  "Joins rows from other datasets"
   ([dataset & args]
   (apply tablecloth.api.dataset/concat dataset args)))
 
 
 (defn concat-copying
+  "Joins rows from other datasets via a copy of data"
   ([dataset & args]
   (apply tablecloth.api.dataset/concat-copying dataset args)))
 
@@ -204,6 +267,7 @@
 
 
 (defn cross-join
+  "Cross product from selected columns"
   ([ds-left ds-right]
   (tablecloth.api.join-concat-ds/cross-join ds-left ds-right))
   ([ds-left ds-right columns-selector]
@@ -340,6 +404,13 @@
 
 
 (defn fill-range-replace
+  "Fill missing up with lacking values. Accepts
+  * dataset
+  * column name
+  * expected step (max-span, milliseconds in case of datetime column)
+  * (optional) missing-strategy - how to replace missing, default :down (set to nil if none)
+  * (optional) missing-value - optional value for replace missing
+"
   ([ds colname max-span]
   (tablecloth.api.missing/fill-range-replace ds colname max-span))
   ([ds colname max-span missing-strategy]
@@ -349,6 +420,7 @@
 
 
 (defn first
+  "First row"
   ([ds]
   (tablecloth.api.rows/first ds)))
 
@@ -365,10 +437,17 @@
 
 
 (defn full-join
+  "Join keeping all rows"
   ([ds-left ds-right columns-selector]
   (tablecloth.api.join-concat-ds/full-join ds-left ds-right columns-selector))
   ([ds-left ds-right columns-selector options]
   (tablecloth.api.join-concat-ds/full-join ds-left ds-right columns-selector options)))
+
+
+(defn get-entry
+  "Returns a single value from given column and row"
+  ([ds column row]
+  (tablecloth.api.dataset/get-entry ds column row)))
 
 
 (defn group-by
@@ -409,6 +488,7 @@
 
 
 (defn groups->seq
+  "Convert grouped dataset to seq of the groups"
   ([ds]
   (tablecloth.api.group-by/groups->seq ds)))
 
@@ -419,6 +499,7 @@
 
 
 (defn head
+  "First n rows (default 5)"
   ([ds]
   (tablecloth.api.rows/head ds))
   ([ds n]
@@ -426,6 +507,9 @@
 
 
 (defn info
+  "Returns a statistcial information about the columns of a dataset.
+  `result-type ` can be :descriptive or :columns
+  "
   ([ds]
   (tablecloth.api.dataset/info ds))
   ([ds result-type]
@@ -447,6 +531,19 @@
 
 
 (defn join-columns
+  "Join clumns of dataset. Accepts:
+dataset
+column selector (as in select-columns)
+options
+  `:separator` (default -)
+  `:drop-columns?` - whether to drop source columns or not (default true)
+  `:result-type`
+     `:map` - packs data into map
+     `:seq` - packs data into sequence
+     `:string` - join strings with separator (default)
+     or custom function which gets row as a vector
+  `:missing-subst` - substitution for missing value
+  "
   ([ds target-column columns-selector]
   (tablecloth.api.join-separate/join-columns ds target-column columns-selector))
   ([ds target-column columns-selector conf]
@@ -454,6 +551,7 @@
 
 
 (defn last
+  "Last row"
   ([ds]
   (tablecloth.api.rows/last ds)))
 
@@ -473,6 +571,7 @@
 
 
 (defn map-columns
+  "Map over rows using a map function. The arity should match the columns selected."
   ([ds column-name map-fn]
   (tablecloth.api.columns/map-columns ds column-name map-fn))
   ([ds column-name columns-selector map-fn]
@@ -516,6 +615,23 @@
 
 
 (defn pivot->wider
+  "Converts columns to rows. Arguments:
+  * dataset
+  * columns selector
+  * options:
+    `:target-columns` - names of the columns created or columns pattern (see below) (default: :$column)
+    `:value-column-name` - name of the column for values (default: :$value)
+    `:splitter` - string, regular expression or function which splits source column names into data
+    `:drop-missing?` - remove rows with missing? (default: true)
+    `:datatypes` - map of target columns data types
+    `:coerce-to-number` - try to convert extracted values to numbers if possible (default: true)
+
+  * target-columns - can be:
+
+    * column name - source columns names are put there as a data
+    * column names as seqence - source columns names after split are put separately into :target-columns as data
+    * pattern - is a sequence of names, where some of the names are nil. nil is replaced by a name taken from splitter and such column is used for values.
+  "
   ([ds columns-selector value-columns]
   (tablecloth.api.reshape/pivot->wider ds columns-selector value-columns))
   ([ds columns-selector value-columns options]
@@ -523,6 +639,9 @@
 
 
 (defn print-dataset
+  "Prints dataset into console. For options see
+  tech.v3.dataset.print/dataset-data->str
+"
   ([ds]
   (tablecloth.api.dataset/print-dataset ds))
   ([ds options]
@@ -530,6 +649,7 @@
 
 
 (defn process-group-data
+  "Internal: The passed-in function is applied on all groups"
   ([ds f]
   (tablecloth.api.utils/process-group-data ds f))
   ([ds f parallel?]
@@ -537,6 +657,7 @@
 
 
 (defn rand-nth
+  "Returns single random row"
   ([ds]
   (tablecloth.api.rows/rand-nth ds))
   ([ds options]
@@ -544,6 +665,7 @@
 
 
 (defn random
+  "Returns (n) random rows with repetition"
   ([ds]
   (tablecloth.api.rows/random ds))
   ([ds n]
@@ -572,6 +694,27 @@
 
 
 (defn replace-missing
+  "Replaces missing values. Accepts
+
+  * dataset
+  * column selector, default: :all
+  * strategy, default: :nearest
+  * value (optional)
+  * single value
+  * sequence of values (cycled)
+  * function, applied on column(s) with stripped missings
+
+  Strategies are:
+
+  `:value` - replace with given value
+  `:up` - copy values up
+  `:down` - copy values down
+  `:updown` - copy values up and then down for missing values at the end
+  `:downup` - copy values down and then up for missing values at the beginning
+  `:mid` or `:nearest` - copy values around known values
+  `:midpoint` - use average value from previous and next non-missing
+  `:lerp` - trying to lineary approximate values, works for numbers and datetime, otherwise applies :nearest. For numbers always results in float datatype.
+  "
   ([ds]
   (tablecloth.api.missing/replace-missing ds))
   ([ds strategy]
@@ -660,6 +803,8 @@
 
 
 (defn separate-column
+  ([ds column]
+  (tablecloth.api.join-separate/separate-column ds column))
   ([ds column separator]
   (tablecloth.api.join-separate/separate-column ds column separator))
   ([ds column target-columns separator]
@@ -680,6 +825,7 @@
 
 
 (defn shuffle
+  "Shuffle dataset (with seed)"
   ([ds]
   (tablecloth.api.rows/shuffle ds))
   ([ds options]
@@ -736,6 +882,7 @@
 
 
 (defn tail
+  "Last n rows (default 5)"
   ([ds]
   (tablecloth.api.rows/tail ds))
   ([ds n]
@@ -760,6 +907,14 @@
 
 
 (defn unique-by
+  "Remove rows which contains the same data
+  `column-selector` Select columns for uniqueness
+  `strategy` There are 4 strategies defined to handle duplicates
+
+    `:first` - select first row (default)
+    `:last` - select last row
+    `:random` - select random row
+    any function - apply function to a columns which are subject of uniqueness"
   ([ds]
   (tablecloth.api.unique-by/unique-by ds))
   ([ds columns-selector]